Refactored AddDataSourceCallback and DataSource out of AIM
